pyproject.toml
[project]
name = "spotify-bigdata-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mypy>=1.19.0",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "pydantic>=2.12.5",
    "pymongo>=4.15.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.1",
]
